requirements.txt
python-vlc
pygame
kivy
Cython
pytaglib